Full Reject, Manage Document and Ocon9x Test (#105)
* Update dependency @axe-core/playwright to v4.9.1

* Bump ansi-regex from 3.0.0 to 3.0.1

Bumps [ansi-regex](https://github.com/chalk/ansi-regex) from 3.0.0 to 3.0.1.
- [Release notes](https://github.com/chalk/ansi-regex/releases)
- [Commits](https://github.com/chalk/ansi-regex/compare/v3.0.0...v3.0.1)

---
updated-dependencies:
- dependency-name: ansi-regex
  dependency-type: indirect
...


* Bump minimist from 1.2.5 to 1.2.8

Bumps [minimist](https://github.com/minimistjs/minimist) from 1.2.5 to 1.2.8.
- [Changelog](https://github.com/minimistjs/minimist/blob/main/CHANGELOG.md)
- [Commits](https://github.com/minimistjs/minimist/compare/v1.2.5...v1.2.8)

---
updated-dependencies:
- dependency-name: minimist
  dependency-type: indirect
...


* Bump minimatch from 3.0.4 to 3.1.2

Bumps [minimatch](https://github.com/isaacs/minimatch) from 3.0.4 to 3.1.2.
- [Changelog](https://github.com/isaacs/minimatch/blob/main/changelog.md)
- [Commits](https://github.com/isaacs/minimatch/compare/v3.0.4...v3.1.2)

---
updated-dependencies:
- dependency-name: minimatch
  dependency-type: indirect
...


* Update dependency axios to v1.7.2

* Update dependency eslint to v8.57.0

* Update dependency randomstring to v1.3.0

* Update playwright monorepo to v1.44.1

* Bump braces from 3.0.2 to 3.0.3

Bumps [braces](https://github.com/micromatch/braces) from 3.0.2 to 3.0.3.
- [Changelog](https://github.com/micromatch/braces/blob/master/CHANGELOG.md)
- [Commits](https://github.com/micromatch/braces/compare/3.0.2...3.0.3)

---
updated-dependencies:
- dependency-name: braces
  dependency-type: indirect
...


* Update base-page.ts

* Update filePaths.ts

* Update date-helper.ts

* Update caseworker-events-factory.ts

* Update link-letter-holder-page.ts

* Create manage-documents-content.ts

* Create manage-documents-page.ts

* Update caseworker-events-steps.ts

* Update FullReject.spec.ts

* Create manageDocuments.spec.ts

* Create fileupload.txt

* Create test.pdf

* Update base-page.ts

* Update dependency puppeteer to v22.11.2

* Update dependency codeceptjs to v3.6.4

* Update dependency @types/node to v20.14.6

* Disabling eslint

* Update package.json

* fixed managed doc tests and re enabled lint checks

* Add more to manage documents test

* Create claim defendant as organisation and full reject test

* ocon9x test

* eslint fix

* fixed full reject test

---------

+++ playwright/global/teardown.ts
@@ -1,3 +1,4 @@
+import config from '../../src/test/config';
 import filePaths from '../config/filePaths';
 import urls from '../config/urls';
 import FileType from '../enums/file-type';
@@ -24,10 +25,12 @@ const deleteUsers = async (userType: UserType) => {
 };
 
 const globalTeardown = async () => {
-  await deleteUsers(UserType.CLAIMANT);
-  await deleteUsers(UserType.DEFENDANT);
-  CitizenUsersHelper.deleteAllUsersState();
-  FileSystemHelper.delete(`${filePaths.userCookies}/`);
+  if(config.skipAuthSetup) {
+    await deleteUsers(UserType.CLAIMANT);
+    await deleteUsers(UserType.DEFENDANT);
+    CitizenUsersHelper.deleteAllUsersState();
+    FileSystemHelper.delete(`${filePaths.userCookies}/`);
+  }
 };
 
 export default globalTeardown;
